reorganized antenna testing files and added collected data
