+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_air_acrossroom.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xcf\x00\x00Axxxx\xd1\x00\x00Axxxx',b'xxxx\xcf\x00\x00Axxxx\xd1\x00\x00Axxxx',48,50
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd2\x00\x00Axxxx\xd3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd2\x00\x00Axxxx\xd3\x00\x00Axxxx',44,42
+0,915,b'xxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',43,43
+0,915,b'xxxx\xd7\x00\x00Axxxx',b'xxxx\xd7\x00\x00Axxxx',41,40
+0,915,b'xxxx\xd5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',44,43
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',43,43
+0,915,b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',42,42
+0,915,b'xxxx\xd6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',43,43
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',43,44
+0,915,b'xxxx\xd5\x00\x00Axxxx\xd6\x00\x00Axxxx',b'xxxx\xd5\x00\x00Axxxx\xd6\x00\x00Axxxx',42,44
+0,915,b'xxxx\xd6\x00\x00Axxxx',b'xxxx\xd6\x00\x00Axxxx',42,43
+0,915,b'xxxx\xd4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',44,45
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',44,48
+0,915,b'xxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',42,42
+0,915,b'xxxx\xd4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',45,46
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd3\x00\x00Axxxx',44,43
+0,915,b'xxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',44,44
+0,915,b'xxxx\xd4\x00\x00Axxxx',b'xxxx\xd4\x00\x00Axxxx',45,45
+0,915,b'xxxx\xd3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',45,43
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd4\x00\x00Axxxx',45,45
+0,915,b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',42,41
+0,915,b'xxxx\xd6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',42,42
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd3\x00\x00Axxxx\xd3\x00\x00Axxxx',45,49
+0,915,b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',b'xxxx\xd5\x00\x00Axxxx\xd5\x00\x00Axxxx',43,43
