+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_air_original.csv
@@ -1,0 +1,18 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xdf\x00\x00Axxxx\xde\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx\xde\x00\x00Axxxx',36,38
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdd\x00\x00Axxxx\xde\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdd\x00\x00Axxxx\xde\x00\x00Axxxx',34,34
+0,915,b'xxxx\xd7\x00\x00Axxxx\xda\x00\x00Axxxx',b'xxxx\xd7\x00\x00Axxxx\xda\x00\x00Axxxx',35,35
+0,915,b'xxxx\xdd\x00\x00Axxxx',b'xxxx\xdd\x00\x00Axxxx',35,33
+0,915,b'xxxx\xd8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',35,36
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdd\x00\x00Axxxx\xde\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdd\x00\x00Axxxx\xde\x00\x00Axxxx',33,33
+0,915,b'xxxx\xdf\x00\x00Axxxx\xde\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx\xde\x00\x00Axxxx',33,34
+0,915,b'xxxx\xde\x00\x00Axxxx',b'xxxx\xde\x00\x00Axxxx',32,37
+0,915,b'xxxx\xe0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',33,32
+0,915,b'xxxx\xe0\x00\x00Axxxx\xe0\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx\xe0\x00\x00Axxxx',34,32
+0,915,b'xxxx\xe0\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx',32,34
+0,915,b'xxxx\xe0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',34,32
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdf\x00\x00Axxxx\xe0\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdf\x00\x00Axxxx\xe0\x00\x00Axxxx',32,32
+0,915,b'xxxx\xe0\x00\x00Axxxx\xdf\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx\xdf\x00\x00Axxxx',33,33
+0,915,b'xxxx\xdf\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx',33,32
+0,915,b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',33,34
+0,915,b'xxxx',b'xxxx',33,97
