+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_water.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,,
+0,915,b'xxxx\xed\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\xed\x00\x00Axxxx\xec\x00\x00Axxxx',20,21,,,*Note: The Dipole was only like 1/4 in the water.
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',19,19,,,
+0,915,b'xxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',25,23,,,
+0,915,b'xxxx\xe4\x00\x00Axxxx',b'xxxx\xe4\x00\x00Axxxx',34,21,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',9,18,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',21,21,,,
+0,915,b'xxxx\xe9\x00\x00Axxxx\xe9\x00\x00Axxxx',b'xxxx\xe9\x00\x00Axxxx\xe9\x00\x00Axxxx',22,21,,,
+0,915,b'xxxx\xea\x00\x00Axxxx',b'xxxx\xea\x00\x00Axxxx',22,21,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xea\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xea\x00\x00Axxxx',22,21,,,
+0,915,b'xxxx\xea\x00\x00Axxxx\xea\x00\x00Axxxx',b'xxxx\xea\x00\x00Axxxx\xea\x00\x00Axxxx',22,22,,,
+0,915,b'xxxx\xea\x00\x00Axxxx',b'xxxx\xea\x00\x00Axxxx',21,25,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',22,23,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',21,21,,,
+0,915,b'xxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',21,21,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx',21,23,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',22,21,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xec\x00\x00Axxxx',19,20,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx',21,20,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',22,21,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xeb\x00\x00Axxxx',21,21,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx\xeb\x00\x00Axxxx',20,20,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx',21,21,,,
+0,915,b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xeb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',22,20,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xec\x00\x00Axxxx',20,21,,,
