+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_water2.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,
+0,915,b'xxxx\xe6\x00\x00Axxxx\xe3\x00\x00Axxxx',b'xxxx\xe6\x00\x00Axxxx\xe3\x00\x00Axxxx',26,22,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xee\x00\x00Axxxx\xee\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xee\x00\x00Axxxx\xee\x00\x00Axxxx',11,10,,
+0,915,b'xxxx\xfc\x00\x00Axxxx\xfb\x00\x00Axxxx',b'xxxx\xfc\x00\x00Axxxx\xfb\x00\x00Axxxx',5,6,,
+0,915,b'xxxx\xee\x00\x00Axxxx',b'xxxx\xee\x00\x00Axxxx',20,25,,*The Dipole was about 3/4 in the water
+0,915,b'xxxx\xe4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xe4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',31,33,,*Separation of about 10cm
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd8\x00\x00Axxxx\xd9\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd8\x00\x00Axxxx\xd9\x00\x00Axxxx',40,39,,*RSSI Values are negative
+0,915,b'xxxx\xd8\x00\x00Axxxx\xd9\x00\x00Axxxx',b'xxxx\xd8\x00\x00Axxxx\xd9\x00\x00Axxxx',40,38,,
+0,915,b'xxxx\xd9\x00\x00Axxxx',b'xxxx\xd9\x00\x00Axxxx',38,38,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd1\x00\x00Axxxx\xd2\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd1\x00\x00Axxxx\xd2\x00\x00Axxxx',49,54,,
+0,915,b'xxxx\xd2\x00\x00Axxxx\xd7\x00\x00Axxxx',b'xxxx\xd2\x00\x00Axxxx\xd7\x00\x00Axxxx',45,50,,
+0,915,b'xxxx\xda\x00\x00Axxxx',b'xxxx\xda\x00\x00Axxxx',37,36,,
+0,915,b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',36,32,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xde\x00\x00Axxxx\xdf\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xde\x00\x00Axxxx\xdf\x00\x00Axxxx',33,34,,
+0,915,b'xxxx\xdb\x00\x00Axxxx\xda\x00\x00Axxxx',b'xxxx\xdb\x00\x00Axxxx\xda\x00\x00Axxxx',38,36,,
+0,915,b'xxxx\xdb\x00\x00Axxxx',b'xxxx\xdb\x00\x00Axxxx',37,36,,
+0,915,b'xxxx\xda\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xda\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',37,38,,
+0,915,b'xxxx\xd9\x00\x00Axxxx\xda\x00\x00Axxxx',b'xxxx\xd9\x00\x00Axxxx\xda\x00\x00Axxxx',39,37,,
+0,915,b'xxxx\xdb\x00\x00Axxxx',b'xxxx\xdb\x00\x00Axxxx',38,37,,
+0,915,b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xde\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',35,32,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdf\x00\x00Axxxx\xdf\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdf\x00\x00Axxxx\xdf\x00\x00Axxxx',34,34,,
+0,915,b'xxxx\xdd\x00\x00Axxxx\xdd\x00\x00Axxxx',b'xxxx\xdd\x00\x00Axxxx\xdd\x00\x00Axxxx',35,34,,
+0,915,b'xxxx\xe0\x00\x00Axxxx',b'xxxx\xe0\x00\x00Axxxx',30,29,,
+0,915,b'xxxx\xe2\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xe2\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',31,34,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xe2\x00\x00Axxxx\xe2\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xe2\x00\x00Axxxx\xe2\x00\x00Axxxx',30,30,,
