+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_water3.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,,
+0,915,b'xxxx\xdf\x00\x00Axxxx\xe7\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx\xe7\x00\x00Axxxx',24,23,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xea\x00\x00Axxxx\xec\x00\x00Axxxx',25,25,,,*Separation of 10cm
+0,915,b'xxxx\xf3\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\xf4\x00\x00Axxxx',11,10,,,* Antennas about 3/4 in
+0,915,b'xxxx\xf0\x00\x00Axxxx',b'xxxx\xf0\x00\x00Axxxx',18,22,,,
+0,915,b'xxxx\xe8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xe8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',26,26,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xe6\x00\x00Axxxx\xe5\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xe6\x00\x00Axxxx\xe5\x00\x00Axxxx',31,32,,,
+0,915,b'xxxx\xe1\x00\x00Axxxx\xe3\x00\x00Axxxx',b'xxxx\xe1\x00\x00Axxxx\xe3\x00\x00Axxxx',30,30,,,
+0,915,b'xxxx\xe3\x00\x00Axxxx',b'xxxx\xe3\x00\x00Axxxx',29,29,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xe3\x00\x00Axxxx\xe2\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xe3\x00\x00Axxxx\xe2\x00\x00Axxxx',30,31,,,
+0,915,b'xxxx\xe2\x00\x00Axxxx\xe2\x00\x00Axxxx',b'xxxx\xe2\x00\x00Axxxx\xe2\x00\x00Axxxx',29,31,,,
+0,915,b'xxxx\xe2\x00\x00Axxxx',b'xxxx\xe2\x00\x00Axxxx',30,31,,,
+0,915,b'xxxx\xdf\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',36,36,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdb\x00\x00Axxxx\xdc\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdb\x00\x00Axxxx\xdc\x00\x00Axxxx',39,38,,,
+0,915,b'xxxx\xdb\x00\x00Axxxx\xdd\x00\x00Axxxx',b'xxxx\xdb\x00\x00Axxxx\xdd\x00\x00Axxxx',35,34,,,
+0,915,b'xxxx\xdd\x00\x00Axxxx',b'xxxx\xdd\x00\x00Axxxx',35,34,,,
+0,915,b'xxxx\xdd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xdd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',36,36,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdc\x00\x00Axxxx\xdc\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdc\x00\x00Axxxx\xdc\x00\x00Axxxx',35,38,,,
+0,915,b'xxxx\xdd\x00\x00Axxxx',b'xxxx\xdd\x00\x00Axxxx',36,38,,,
+0,915,b'xxxx\xd9\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xd9\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',42,44,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd4\x00\x00Axxxx\xd4\x00\x00Axxxx',45,44,,,
+0,915,b'xxxx\xd6\x00\x00Axxxx\xd7\x00\x00Axxxx',b'xxxx\xd6\x00\x00Axxxx\xd7\x00\x00Axxxx',40,40,,,
+0,915,b'xxxx\xd8\x00\x00Axxxx',b'xxxx\xd8\x00\x00Axxxx',39,40,,,
+0,915,b'xxxx\xcd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xcd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',55,48,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xdc\x00\x00Axxxx\xdd\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xdc\x00\x00Axxxx\xdd\x00\x00Axxxx',36,37,,,
