+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_water4.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,,
+0,915,b'xxxx\xfa\x00\x00Axxxx\xfe\x00\x00Axxxx',b'xxxx\xfa\x00\x00Axxxx\xfe\x00\x00Axxxx',6,12,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xeb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xeb\x00\x00Axxxx\xeb\x00\x00Axxxx',32,19,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\xf2\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\xf2\x00\x00Axxxx',34,34,,,*With Cable
+0,915,b'xxxx\xe6\x00\x00Axxxx',b'xxxx\xe6\x00\x00Axxxx',24,30,,,*About 10cm apart
+0,915,b'xxxx\xf1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',17,19,,,*about 3/4 in the water
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xee\x00\x00Axxxx\xf0\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xee\x00\x00Axxxx\xf0\x00\x00Axxxx',19,20,,,
+0,915,b'xxxx\xec\x00\x00Axxxx\xee\x00\x00Axxxx',b'xxxx\xec\x00\x00Axxxx\xee\x00\x00Axxxx',16,13,,,
+0,915,b'xxxx\xf2\x00\x00Axxxx',b'xxxx\xf2\x00\x00Axxxx',14,10,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf5\x00\x00Axxxx\xf5\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf5\x00\x00Axxxx\xf5\x00\x00Axxxx',11,11,,,
+0,915,b'xxxx\xf7\x00\x00Axxxx\xf5\x00\x00Axxxx',b'xxxx\xf7\x00\x00Axxxx\xf5\x00\x00Axxxx',11,12,,,
+0,915,b'xxxx\xf5\x00\x00Axxxx',b'xxxx\xf5\x00\x00Axxxx',10,10,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',14,12,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\xf5\x00\x00Axxxx\xf5\x00\x00Axxxx',b'xxxx\xf5\x00\x00Axxxx\xf5\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx',12,10,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',11,11,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,13,,,
+0,915,b'xxxx\xf1\x00\x00Axxxx',b'xxxx\xf1\x00\x00Axxxx',14,13,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',11,10,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\xf5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',12,15,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf5\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf5\x00\x00Axxxx',11,11,,,
