+++ collected_data/Preliminary_AntennaTests_Jul11/dipole_water5.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,,,
+0,915,b'xxxx\xf0\x00\x00Axxxx\xee\x00\x00Axxxx',b'xxxx\xf0\x00\x00Axxxx\xee\x00\x00Axxxx',18,17,,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xe3\x00\x00Axxxx\xed\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xe3\x00\x00Axxxx\xed\x00\x00Axxxx',17,18,,,,
+0,915,b'xxxx\xf1\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf1\x00\x00Axxxx\xf4\x00\x00Axxxx',9,7,,,,*With Cable
+0,915,b'xxxx\xfa\x00\x00Axxxx',b'xxxx\xfa\x00\x00Axxxx',8,16,,,,*About 10cm apart
+0,915,b'xxxx\xec\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xec\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',17,15,,,,*about 3/4 in the water
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf7\x00\x00Axxxx\xf8\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf7\x00\x00Axxxx\xf8\x00\x00Axxxx',10,10,,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',11,12,,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx',10,11,,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',11,10,,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',12,12,,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx',12,13,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',13,13,,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,12,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,12,,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx',12,16,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',13,13,,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,15,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx',13,16,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',13,12,,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,12,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,13,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx',12,12,,,,
+0,915,b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf3\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',14,13,,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',13,14,,,,
