+++ collected_data/Preliminary_AntennaTests_Jul11/jpole_air.csv
@@ -1,0 +1,18 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',80,80
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',79,79
+0,915,b'xxxx\xb0\x00\x00Axxxx\xae\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx\xae\x00\x00Axxxx',82,80
+0,915,b'xxxx\xb0\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx',79,81
+0,915,b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',79,82
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb1\x00\x00Axxxx\xb2\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb1\x00\x00Axxxx\xb2\x00\x00Axxxx',78,78
+0,915,b'xxxx\xb2\x00\x00Axxxx\xb2\x00\x00Axxxx',b'xxxx\xb2\x00\x00Axxxx\xb2\x00\x00Axxxx',78,77
+0,915,b'xxxx\xb2\x00\x00Axxxx',b'xxxx\xb2\x00\x00Axxxx',78,78
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb1\x00\x00Axxxx\xb2\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb1\x00\x00Axxxx\xb2\x00\x00Axxxx',79,81
+0,915,b'xxxx\xae\x00\x00Axxxx\xae\x00\x00Axxxx',b'xxxx\xae\x00\x00Axxxx\xae\x00\x00Axxxx',79,81
+0,915,b'xxxx\xb2\x00\x00Axxxx',b'xxxx\xb2\x00\x00Axxxx',79,80
+0,915,b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',80,79
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',80,81
+0,915,b'xxxx\xb1\x00\x00Axxxx\xb1\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\xb1\x00\x00Axxxx',79,79
+0,915,b'xxxx\xb1\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx',79,79
+0,915,b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',79,80
+0,915,b'xxxx',b'xxxx',79,106
