+++ collected_data/Preliminary_AntennaTests_Jul11/loop_air.csv
@@ -1,0 +1,18 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xbe\x00\x00Axxxx\xbd\x00\x00Axxxx',b'xxxx\xbe\x00\x00Axxxx\xbd\x00\x00Axxxx',71,66
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbe\x00\x00Axxxx\xb7\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbe\x00\x00Axxxx\xb7\x00\x00Axxxx',69,70
+0,915,b'xxxx\xbc\x00\x00Axxxx\xb6\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\xb6\x00\x00Axxxx',71,73
+0,915,b'xxxx\xba\x00\x00Axxxx',b'xxxx\xba\x00\x00Axxxx',73,69
+0,915,b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',69,67
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbe\x00\x00Axxxx\xba\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbe\x00\x00Axxxx\xba\x00\x00Axxxx',67,65
+0,915,b'xxxx\xbc\x00\x00Axxxx\xb6\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\xb6\x00\x00Axxxx',75,72
+0,915,b'xxxx\xb3\x00\x00Axxxx',b'xxxx\xb3\x00\x00Axxxx',80,93
+0,915,b'xxxx\xaa\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xaa\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',77,76
+0,915,b'xxxx\xb6\x00\x00Axxxx\xbc\x00\x00Axxxx',b'xxxx\xb6\x00\x00Axxxx\xbc\x00\x00Axxxx',63,63
+0,915,b'xxxx\xba\x00\x00Axxxx',b'xxxx\xba\x00\x00Axxxx',68,67
+0,915,b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',69,71
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb9\x00\x00Axxxx\xba\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb9\x00\x00Axxxx\xba\x00\x00Axxxx',70,70
+0,915,b'xxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',69,69
+0,915,b'xxxx\xbb\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx',68,71
+0,915,b'xxxx\xbe\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbe\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',68,70
+0,915,b'xxxx',b'xxxx',64,104
