+++ collected_data/Preliminary_AntennaTests_Jul11/loop_air2_test.csv
@@ -1,0 +1,18 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xba\x00\x00Axxxx\xb8\x00\x00Axxxx',b'xxxx\xba\x00\x00Axxxx\xb8\x00\x00Axxxx',72,72
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xb9\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xb9\x00\x00Axxxx',69,68
+0,915,b'xxxx\xbf\x00\x00Axxxx\xbd\x00\x00Axxxx',b'xxxx\xbf\x00\x00Axxxx\xbd\x00\x00Axxxx',66,66
+0,915,b'xxxx\xbc\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx',66,67
+0,915,b'xxxx\xbd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',68,69
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',69,69
+0,915,b'xxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',69,69
+0,915,b'xxxx\xba\x00\x00Axxxx',b'xxxx\xba\x00\x00Axxxx',70,69
+0,915,b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',69,69
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbb\x00\x00Axxxx',68,68
+0,915,b'xxxx\xbc\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx',68,68
+0,915,b'xxxx\xbc\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',69,68
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbc\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbb\x00\x00Axxxx\xbc\x00\x00Axxxx',68,68
+0,915,b'xxxx\xbc\x00\x00Axxxx\xbc\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\xbc\x00\x00Axxxx',68,68
+0,915,b'xxxx\xbc\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx',68,68
+0,915,b'xxxx\xbc\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',68,68
+0,915,b'xxxx',b'xxxx',69,105
