+++ collected_data/Preliminary_AntennaTests_Jul11/loop_air3_test.csv
@@ -1,0 +1,18 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xbe\x00\x00Axxxx\xbf\x00\x00Axxxx',b'xxxx\xbe\x00\x00Axxxx\xbf\x00\x00Axxxx',65,65
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xbc\x00\x00Axxxx\xbc\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xbc\x00\x00Axxxx\xbc\x00\x00Axxxx',72,71
+0,915,b'xxxx\xbc\x00\x00Axxxx\xbb\x00\x00Axxxx',b'xxxx\xbc\x00\x00Axxxx\xbb\x00\x00Axxxx',74,80
+0,915,b'xxxx\xb2\x00\x00Axxxx',b'xxxx\xb2\x00\x00Axxxx',80,80
+0,915,b'xxxx\xaf\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xaf\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',84,85
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xab\x00\x00Axxxx\xad\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xab\x00\x00Axxxx\xad\x00\x00Axxxx',82,82
+0,915,b'xxxx\xb1\x00\x00Axxxx\xb4\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\xb4\x00\x00Axxxx',76,76
+0,915,b'xxxx\xb1\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx',80,83
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',80,81
+0,915,b'xxxx\xb1\x00\x00Axxxx\xb1\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\xb1\x00\x00Axxxx',79,79
+0,915,b'xxxx\xb0\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx',80,79
+0,915,b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xb1\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',80,79
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xb0\x00\x00Axxxx\xb1\x00\x00Axxxx',79,79
+0,915,b'xxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx\xb0\x00\x00Axxxx',80,79
+0,915,b'xxxx\xb0\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx',80,80
+0,915,b'xxxx\xb0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xb0\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',80,80
+0,915,b'xxxx',b'xxxx',80,104
