+++ collected_data/Preliminary_AntennaTests_Jul11/loop_water.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI,,,
+0,915,b'xxxx\xe7\x00\x00Axxxx\xe3\x00\x00Axxxx',b'xxxx\xe7\x00\x00Axxxx\xe3\x00\x00Axxxx',29,29,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xe4\x00\x00Axxxx\xe8\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xe4\x00\x00Axxxx\xe8\x00\x00Axxxx',24,22,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\xf1\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\xf1\x00\x00Axxxx',20,20,,,
+0,915,b'xxxx\xf9\x00\x00Axxxx',b'xxxx\xf9\x00\x00Axxxx',5,5,,,
+0,915,b'xxxx\xfa\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xfa\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',7,6,,,*loop antenna
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xf9\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xf9\x00\x00Axxxx',7,8,,,*with cable
+0,915,b'xxxx\xf8\x00\x00Axxxx\xf8\x00\x00Axxxx',b'xxxx\xf8\x00\x00Axxxx\xf8\x00\x00Axxxx',8,5,,,*about 12cm
+0,915,b'xxxx\xfb\x00\x00Axxxx',b'xxxx\xfb\x00\x00Axxxx',6,5,,,*whole loop in the water with one loop not facing the other loop (radiation patterns not facing each other)
+0,915,b'xxxx\xfb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xfb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',4,3,,,
+0,915,b'xxxx\xfc\x00\x00Axxxx\xfa\x00\x00Axxxx',b'xxxx\xfc\x00\x00Axxxx\xfa\x00\x00Axxxx',7,7,,,
+0,915,b'xxxx\xfa\x00\x00Axxxx',b'xxxx\xfa\x00\x00Axxxx',7,7,,,
+0,915,b'xxxx\xf8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf8\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',9,10,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',10,10,,,
+0,915,b'xxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx\xf6\x00\x00Axxxx',11,11,,,
+0,915,b'xxxx\xf5\x00\x00Axxxx',b'xxxx\xf5\x00\x00Axxxx',11,13,,,
+0,915,b'xxxx\xf5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf5\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',12,11,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',12,12,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',12,12,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',13,12,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf3\x00\x00Axxxx\xf3\x00\x00Axxxx',12,16,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx\xf4\x00\x00Axxxx',13,12,,,
+0,915,b'xxxx\xf4\x00\x00Axxxx',b'xxxx\xf4\x00\x00Axxxx',11,10,,,
+0,915,b'xxxx\xf7\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf7\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',7,6,,,
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xf8\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xf8\x00\x00Axxxx',9,10,,,
