+++ collected_data/Preliminary_AntennaTests_Jul11/loop_water2.csv
@@ -1,0 +1,25 @@
+Address,Frequency,Message,String Message,Packet RSSI,Channel RSSI
+0,915,b'xxxx\xdf\x00\x00Axxxx\xdb\x00\x00Axxxx',b'xxxx\xdf\x00\x00Axxxx\xdb\x00\x00Axxxx',37,38
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xd2\x00\x00Axxxx\xcd\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xd2\x00\x00Axxxx\xcd\x00\x00Axxxx',37,37
+0,915,b'xxxx\xd7\x00\x00Axxxx\xdb\x00\x00Axxxx',b'xxxx\xd7\x00\x00Axxxx\xdb\x00\x00Axxxx',25,20
+0,915,b'xxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',256,256
+0,915,b'xxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx\x00\x00\x00Axxxx',256,1
+0,915,b'xxxx\x00\x00\x00Axxxx',b'xxxx\x00\x00\x00Axxxx',256,1
+0,915,b'xxxx\xff\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xff\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',2,2
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xff\x00\x00Axxxx\xff\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xff\x00\x00Axxxx\xff\x00\x00Axxxx',1,1
+0,915,b'xxxx\xff\x00\x00Axxxx\xff\x00\x00Axxxx',b'xxxx\xff\x00\x00Axxxx\xff\x00\x00Axxxx',2,1
+0,915,b'xxxx\xfe\x00\x00Axxxx',b'xxxx\xfe\x00\x00Axxxx',2,3
+0,915,b'xxxx\xfd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xfd\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',3,5
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xfc\x00\x00Axxxx\xfc\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xfc\x00\x00Axxxx\xfc\x00\x00Axxxx',3,4
+0,915,b'xxxx\xfd\x00\x00Axxxx\xfc\x00\x00Axxxx',b'xxxx\xfd\x00\x00Axxxx\xfc\x00\x00Axxxx',4,5
+0,915,b'xxxx\xfb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xfb\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',7,6
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xfa\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf9\x00\x00Axxxx\xfa\x00\x00Axxxx',7,8
+0,915,b'xxxx\xf8\x00\x00Axxxx\xf6\x00\x00Axxxx',b'xxxx\xf8\x00\x00Axxxx\xf6\x00\x00Axxxx',11,11
+0,915,b'xxxx\xf6\x00\x00Axxxx',b'xxxx\xf6\x00\x00Axxxx',10,12
+0,915,b'xxxx\xf7\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',b'xxxx\xf7\x00\x00Axxxx\x00\x00A\x00\x00Axxxx',10,12
+0,915,b'xxxx\x00\x00A\x00\x00Axxxx\xf5\x00\x00Axxxx\xf6\x00\x00Axxxx',b'xxxx\x00\x00A\x00\x00Axxxx\xf5\x00\x00Axxxx\xf6\x00\x00Axxxx',11,10
